+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# 1.3.0
+* Make `choose` only print options once, and instead add `prompt`.
+* Add `names` to `choose`.
+* `conceal` was `true` in `chooseShorthand`; revert to `false`.
+
 # 1.2.0
 * Add password reading options.
 * Pin against Dart 2.